+++ spec/dummy/app/models/author.rb
@@ -3,6 +3,7 @@ class Author < ActiveRecord::Base
   include Storext
   store_attributes :data do
     name String
+    title String, default: "Dr."
   end
 
 end