+++ spec/dummy/app/models/pocket_book.rb
@@ -1,0 +1,5 @@
+class PocketBook < Book
+
+  store_attribute :data, :soft, Boolean, default: true
+
+end