+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBA
+
+- Fix: errors that show up when `include Storext` is in parent
+
 # 0.1.1
 
 - Fix: defaults are now set when an object is created